// src/generated/prisma/enums.ts
export const AuditAction = {
  LOGIN: 'LOGIN',
  LOGOUT: 'LOGOUT',
  CALL_CREATED: 'CALL_CREATED',
  CALL_UPDATED: 'CALL_UPDATED',
  CALL_DELETED: 'CALL_DELETED',
  CALL_IMPORTED: 'CALL_IMPORTED',
  REMINDER_CREATED: 'REMINDER_CREATED',
  REMINDER_UPDATED: 'REMINDER_UPDATED',
  REMINDER_DELETED: 'REMINDER_DELETED',
  SETTINGS_UPDATED: 'SETTINGS_UPDATED',
  USER_CREATED: 'USER_CREATED',
  USER_UPDATED: 'USER_UPDATED',
  USER_ACTIVATED: 'USER_ACTIVATED',
  USER_DEACTIVATED: 'USER_DEACTIVATED',
  PASSWORD_RESET: 'PASSWORD_RESET',
  CLIENT_CREATED: 'CLIENT_CREATED',
  CLIENT_UPDATED: 'CLIENT_UPDATED',
  CLIENT_DELETED: 'CLIENT_DELETED',
  EVENT_CREATED: 'EVENT_CREATED',
  EVENT_UPDATED: 'EVENT_UPDATED',
  EVENT_DELETED: 'EVENT_DELETED',
  APPOINTMENT_STATUS_CHANGED: 'APPOINTMENT_STATUS_CHANGED',
  ROLE_CREATED: 'ROLE_CREATED',
  ROLE_UPDATED: 'ROLE_UPDATED',
  ROLE_DELETED: 'ROLE_DELETED',
  CONFIGURABLE_LIST_ITEM_CREATED: 'CONFIGURABLE_LIST_ITEM_CREATED',
  CONFIGURABLE_LIST_ITEM_UPDATED: 'CONFIGURABLE_LIST_ITEM_UPDATED',
  CUSTOM_FIELD_CREATED: 'CUSTOM_FIELD_CREATED',
  CUSTOM_FIELD_UPDATED: 'CUSTOM_FIELD_UPDATED',
  CUSTOM_FIELD_DELETED: 'CUSTOM_FIELD_DELETED',
  EMAIL_TEMPLATE_CREATED: 'EMAIL_TEMPLATE_CREATED',
  EMAIL_TEMPLATE_UPDATED: 'EMAIL_TEMPLATE_UPDATED',
  EMAIL_SENT: 'EMAIL_SENT',
  ORGANIZATION_CREATED: 'ORGANIZATION_CREATED',
  ORGANIZATION_UPDATED: 'ORGANIZATION_UPDATED',
  ORGANIZATION_DEACTIVATED: 'ORGANIZATION_DEACTIVATED'
} as const

export type AuditAction = (typeof AuditAction)[keyof typeof AuditAction]